Refactoring of high-level request handler class structure
