+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/GempHttpRequest.java
@@ -1,0 +1,136 @@
+package com.gempukku.stccg.async;
+
+import com.gempukku.stccg.database.IpBanDAO;
+import com.gempukku.stccg.database.PlayerDAO;
+import com.gempukku.stccg.database.User;
+import com.gempukku.stccg.service.LoggedUserHolder;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.handler.codec.http.*;
+import io.netty.handler.codec.http.cookie.Cookie;
+import io.netty.handler.codec.http.cookie.ServerCookieDecoder;
+import io.netty.handler.codec.http.multipart.Attribute;
+import io.netty.handler.codec.http.multipart.HttpPostRequestDecoder;
+import io.netty.handler.codec.http.multipart.InterfaceHttpData;
+import io.netty.handler.codec.http.multipart.InterfaceHttpPostRequestDecoder;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.InetSocketAddress;
+import java.util.*;
+
+import static io.netty.handler.codec.http.HttpHeaderNames.COOKIE;
+
+public class GempHttpRequest {
+
+    private final FullHttpRequest _request;
+    private final LoggedUserHolder _loggedUserHolder;
+    private final String _ipAddress;
+    private final boolean _ipBanned;
+    private final User _user;
+
+    public GempHttpRequest(FullHttpRequest request, ChannelHandlerContext context, ServerObjects serverObjects) {
+        _request = request;
+        _loggedUserHolder = serverObjects.getLoggedUserHolder();
+        _ipAddress = Objects.requireNonNullElse(request.headers().get("X-Forwarded-For"),
+                ((InetSocketAddress) context.channel().remoteAddress()).getAddress().getHostAddress()
+        );
+        _ipBanned = determineIfIpIsBanned(serverObjects);
+        _user = identifyUser(serverObjects);
+    }
+
+    private boolean determineIfIpIsBanned(ServerObjects serverObjects) {
+        IpBanDAO ipBanDAO = serverObjects.getIpBanDAO();
+        if (ipBanDAO.getIpBans().contains(_ipAddress))
+            return true;
+        for (String bannedRange : ipBanDAO.getIpPrefixBans()) {
+            if (_ipAddress.startsWith(bannedRange))
+                return true;
+        }
+        return false;
+    }
+
+
+    private User identifyUser(ServerObjects serverObjects) {
+        ServerCookieDecoder cookieDecoder = ServerCookieDecoder.STRICT;
+        String cookieHeader = _request.headers().get(COOKIE);
+        String userName = "";
+        PlayerDAO playerDAO = serverObjects.getPlayerDAO();
+
+        if (cookieHeader != null) {
+            Set<Cookie> cookies = cookieDecoder.decode(cookieHeader);
+            for (Cookie cookie : cookies) {
+                if ("loggedUser".equals(cookie.name())) {
+                    String value = cookie.value();
+                    if (value != null) {
+                        userName = _loggedUserHolder.getLoggedUserNew(value);
+                    }
+                }
+            }
+        }
+        if (!userName.isEmpty()) {
+            return playerDAO.getPlayer(userName);
+        } else {
+            return null;
+        }
+    }
+
+    public boolean bannedIp() {
+        return _ipBanned;
+    }
+
+    public String ip() {
+        return _ipAddress;
+    }
+
+    public String uri() {
+        return _request.uri();
+    }
+
+    public User user() throws HttpProcessingException {
+        if (_user == null) {
+            throw new HttpProcessingException(HttpURLConnection.HTTP_UNAUTHORIZED); // 401
+        } else {
+            return _user;
+        }
+    }
+
+    Map<String, String> parameters() throws IOException {
+        Map<String, String> result = new HashMap<>();
+        if (_request.method() == HttpMethod.POST) {
+            InterfaceHttpPostRequestDecoder postDecoder = new HttpPostRequestDecoder(_request);
+            try {
+                for (InterfaceHttpData data : postDecoder.getBodyHttpDatas()) {
+                    if (data instanceof Attribute attribute) {
+                        result.put(attribute.getName(), attribute.getValue());
+                    }
+                }
+            } finally {
+                postDecoder.destroy();
+            }
+        } else if (_request.method() == HttpMethod.GET) {
+            QueryStringDecoder queryDecoder = new QueryStringDecoder(_request.uri());
+            for (Map.Entry<String, List<String>> entry : queryDecoder.parameters().entrySet()) {
+                if (entry.getValue() != null && !entry.getValue().isEmpty() && !entry.getKey().equals("_")) {
+                    result.put(entry.getKey(), entry.getValue().getFirst());
+                }
+            }
+        }
+        return result;
+    }
+
+    public HttpRequest getRequest() {
+        return _request;
+    }
+
+    public String uriWithoutParameters() {
+        String uri = _request.uri();
+        if (uri.contains("?"))
+            uri = uri.substring(0, uri.indexOf('?'));
+        return uri;
+    }
+
+    public String cookieHeader() {
+        String result = _request.headers().get(HttpHeaderNames.COOKIE);
+        return (result == null) ? "" : result;
+    }
+}
