+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/LeagueRequestHandler.java
@@ -1,6 +1,7 @@
 package com.gempukku.stccg.async.handler;
 
 import com.gempukku.stccg.DateUtils;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.HttpProcessingException;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.competitive.PlayerStanding;
@@ -28,21 +29,22 @@ public class LeagueRequestHandler extends DefaultServerRequestHandler implements
     private final LeagueService _leagueService;
     private final FormatLibrary _formatLibrary;
 
-    LeagueRequestHandler(ServerObjects objects) {
+    public LeagueRequestHandler(ServerObjects objects) {
         super(objects);
         _leagueService = objects.getLeagueService();
         _formatLibrary = objects.getFormatLibrary();
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp)
+    public final void handleRequest(String uri, GempHttpRequest gempRequest, ResponseWriter responseWriter)
             throws Exception {
+        HttpRequest request = gempRequest.getRequest();
         if (uri.isEmpty() && request.method() == HttpMethod.GET) {
             getNonExpiredLeagues(responseWriter);
         } else if (uri.startsWith("/") && request.method() == HttpMethod.GET) {
             getLeagueInformation(request, uri.substring(1), responseWriter);
         } else if (uri.startsWith("/") && request.method() == HttpMethod.POST) {
-            joinLeague(request, uri.substring(1), responseWriter, remoteIp);
+            joinLeague(request, uri.substring(1), responseWriter, gempRequest.ip());
         } else {
             throw new HttpProcessingException(HttpURLConnection.HTTP_NOT_FOUND); // 404
         }
