+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/UriRequestHandler.java
@@ -1,12 +1,13 @@
 package com.gempukku.stccg.async.handler;
 
+import com.gempukku.stccg.async.GempHttpRequest;
 import io.netty.handler.codec.http.HttpRequest;
 import org.apache.logging.log4j.Logger;
 
 import java.net.HttpURLConnection;
 
 public interface UriRequestHandler {
-    void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp)
+    void handleRequest(String uri, GempHttpRequest request, ResponseWriter responseWriter)
             throws Exception;
 
     default void logHttpError(Logger log, int code, String uri, Exception exp) {
