+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/game/CancelGameRequestHandler.java
@@ -1,6 +1,7 @@
 package com.gempukku.stccg.async.handler.game;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
 import com.gempukku.stccg.async.handler.UriRequestHandlerNew;
@@ -18,10 +19,9 @@ public class CancelGameRequestHandler extends GameRequestHandlerNew implements U
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
-                                    ServerObjects serverObjects)
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter, ServerObjects serverObjects)
             throws Exception {
-        User resourceOwner = getResourceOwnerSafely(request, serverObjects);
+        User resourceOwner = request.user();
         GameServer gameServer = serverObjects.getGameServer();
         gameServer.cancelGame(resourceOwner, _gameId);
         responseWriter.writeXmlOkResponse();
