+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/game/GameCardInfoRequestHandler.java
@@ -1,6 +1,7 @@
 package com.gempukku.stccg.async.handler.game;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
 import com.gempukku.stccg.async.handler.UriRequestHandlerNew;
@@ -21,7 +22,7 @@ public class GameCardInfoRequestHandler extends GameRequestHandlerNew implements
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter,
                                     ServerObjects serverObjects)
             throws Exception {
             // getGameById throws 404 error if not found
