+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/game/UpdateGameStateRequestHandler.java
@@ -1,6 +1,7 @@
 package com.gempukku.stccg.async.handler.game;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.LongPollingResource;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
@@ -25,10 +26,9 @@ public class UpdateGameStateRequestHandler extends GameRequestHandlerNew impleme
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
-                                    ServerObjects serverObjects)
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter, ServerObjects serverObjects)
             throws Exception {
-        User resourceOwner = getResourceOwnerSafely(request, serverObjects);
+        User resourceOwner = request.user();
         GameServer gameServer = serverObjects.getGameServer();
         gameServer.setPlayerAutoPassSettings(resourceOwner, _gameId, getAutoPassPhases(request));
