+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/hall/GetErrataRequestHandler.java
@@ -2,6 +2,7 @@ package com.gempukku.stccg.async.handler.hall;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
 import com.gempukku.stccg.async.handler.UriRequestHandlerNew;
@@ -13,8 +14,7 @@ import io.netty.handler.codec.http.HttpRequest;
 
 public class GetErrataRequestHandler implements UriRequestHandlerNew {
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
-                                    ServerObjects serverObjects)
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter, ServerObjects serverObjects)
             throws Exception {
         CardBlueprintLibrary library = serverObjects.getCardBlueprintLibrary();
         String jsonString = new ObjectMapper().writeValueAsString(library.getErrata());
