+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/hall/JoinQueueRequestHandler.java
@@ -1,6 +1,7 @@
 package com.gempukku.stccg.async.handler.hall;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
 import com.gempukku.stccg.async.handler.UriRequestHandlerNew;
@@ -28,10 +29,9 @@ public class JoinQueueRequestHandler implements UriRequestHandlerNew {
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
-                                    ServerObjects serverObjects)
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter, ServerObjects serverObjects)
             throws Exception {
-        User resourceOwner = getResourceOwnerSafely(request, serverObjects);
+        User resourceOwner = request.user();
         HallServer hallServer = serverObjects.getHallServer();
         try {
             hallServer.joinQueue(_queueId, resourceOwner, _deckName);
