+++ gemp-module/gemp-stccg-server/src/main/java/com/gempukku/stccg/async/handler/login/RegisterRequestHandler.java
@@ -2,6 +2,7 @@ package com.gempukku.stccg.async.handler.login;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.async.GempHttpRequest;
 import com.gempukku.stccg.async.HttpProcessingException;
 import com.gempukku.stccg.async.ServerObjects;
 import com.gempukku.stccg.async.handler.ResponseWriter;
@@ -32,12 +33,12 @@ public class RegisterRequestHandler implements UriRequestHandlerNew {
     }
 
     @Override
-    public final void handleRequest(String uri, HttpRequest request, ResponseWriter responseWriter, String remoteIp,
+    public final void handleRequest(GempHttpRequest request, ResponseWriter responseWriter,
                                     ServerObjects serverObjects)
             throws Exception {
         try {
-            if (serverObjects.getPlayerDAO().registerUser(_userId, _password, remoteIp)) {
-                Map<String, String> headers = logUserReturningHeaders(remoteIp, _userId, serverObjects);
+            if (serverObjects.getPlayerDAO().registerUser(_userId, _password, request.ip())) {
+                Map<String, String> headers = logUserReturningHeaders(request.ip(), _userId, serverObjects);
                 responseWriter.writeEmptyXmlResponseWithHeaders(headers);
             } else
                 throw new HttpProcessingException(HttpURLConnection.HTTP_CONFLICT); // 409
